# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/chat":{"get":{"summary":"Chat","operationId":"chat_chat_get","parameters":[{"name":"userid","in":"query","required":true,"schema":{"type":"string","description":"用户ID","title":"Userid"},"description":"用户ID"},{"name":"sessionid","in":"query","required":true,"schema":{"type":"string","description":"会话ID","title":"Sessionid"},"description":"会话ID"},{"name":"user_msg","in":"query","required":true,"schema":{"type":"string","description":"用户消息","title":"User Msg"},"description":"用户消息"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/load_history":{"get":{"summary":"Load History","operationId":"load_history_load_history_get","parameters":[{"name":"userid","in":"query","required":true,"schema":{"type":"string","description":"用户ID","title":"Userid"},"description":"用户ID"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/load_specific_session":{"get":{"summary":"Load Specific Session","operationId":"load_specific_session_load_specific_session_get","parameters":[{"name":"userid","in":"query","required":true,"schema":{"type":"string","description":"用户ID","title":"Userid"},"description":"用户ID"},{"name":"sessionid","in":"query","required":true,"schema":{"type":"string","description":"会话ID","title":"Sessionid"},"description":"会话ID"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/create_user":{"get":{"summary":"Create User","operationId":"create_user_create_user_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/create_new_chat":{"get":{"summary":"Create New Chat","operationId":"create_new_chat_create_new_chat_get","parameters":[{"name":"userid","in":"query","required":true,"schema":{"type":"string","description":"用户ID","title":"Userid"},"description":"用户ID"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat/delete_session":{"get":{"summary":"Delete Session","operationId":"delete_session_chat_delete_session_get","parameters":[{"name":"userid","in":"query","required":true,"schema":{"type":"string","description":"用户ID","title":"Userid"},"description":"用户ID"},{"name":"sessionid","in":"query","required":true,"schema":{"type":"string","description":"会话ID","title":"Sessionid"},"description":"会话ID"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat/save_usermsg":{"get":{"summary":"Save Usermsg","operationId":"save_usermsg_chat_save_usermsg_get","parameters":[{"name":"userid","in":"query","required":true,"schema":{"type":"string","description":"用户ID","title":"Userid"},"description":"用户ID"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}